refactor: extract log streaming out of server.go
